Small refact also makes enter work on forms
