+++ sdk-app/src/components/button/button.tsx
@@ -1,28 +1,19 @@
-import { CSSProperties, FC, useCallback, useRef, useState } from "react"
+import { CSSProperties, FC } from "react"
 import './button.scss'
 
 export type ButtonProps = {
     onClick?: () => void
     className?: string
-    label?: string
-    hoverLabel?: string
-    hideArrow?: boolean
     disabled?: boolean
     style?: CSSProperties
+    label: string
 }
 
-export const Button: FC<ButtonProps> = ({ style ,disabled = false ,onClick, label = 'Submit', className, hoverLabel = 'Next', hideArrow = false }) => {
-    const ref = useRef<HTMLButtonElement>(null)
-    const [innerHtml, setInnerHtml] = useState(label)
+export const Button: FC<ButtonProps> = ({ style, disabled = false, onClick, className, label }) => {
 
-    const innerHtmlSetter = useCallback((label: string) =>() => {
-        setInnerHtml(label)
-    }, [])
-
-    return <button style={style} disabled={disabled} ref={ref} onMouseEnter={innerHtmlSetter(hoverLabel)} onMouseLeave={innerHtmlSetter(label)}  className={`kyc-button ${className}`} onClick={onClick}>
+    return <button style={style} disabled={disabled} className={`kyc-button ${className}`} onClick={onClick}>
         <span>
-            {innerHtml}
+            {label}
         </span>
-        { !hideArrow && <i className="material-icons arrow"> arrow_forward </i>}
     </button>
 }
