+++ sdk-app/src/components/submitButton/button.scss
@@ -1,0 +1,225 @@
+@import "../../style/variables.scss";
+
+button.kyc-button {
+  display: inline-flex;
+  justify-content: space-between;
+  padding-left: 1.5rem;
+  align-items: center;
+  outline: none;
+  border: 1px solid transparent;
+  border-radius: $--border-radius-light;
+  cursor: pointer;
+  height: $--button-height;
+
+  &[disabled] {
+    background: $--inactive!important;
+    pointer-events: none!important;
+  }
+
+  &.centered {
+    justify-content: center;
+    padding-left: 0;
+    padding-right: 0;
+  }
+
+  span {
+    font-family: $--light-font;
+    font-weight: 200;
+    font-size: 18px;
+    margin-left: 0;
+  }
+
+  .arrow {
+    margin-left: 1rem;
+  }
+
+  &:hover {
+    .nav {
+      width: auto;
+      opacity: 1;
+      visibility: visible;
+      transform: translateX(0rem);
+    }
+  }
+
+  &.full-width {
+    display: flex;
+    width: 100%;
+  }
+
+  &.transparent {
+    color: $--kyc-normal-blue;
+    background: transparent;
+
+    &:hover {
+      color: $--kyc-normal-blue-35;
+    }
+  }
+
+  &.white {
+    background: white;
+    box-shadow: 0 0 0 0 rgba(white, 0);
+
+    span,
+    .material-icons {
+      color: $--kyc-dark-purple;
+    }
+
+    &:hover {
+      background: $--kyc-normal-blue;
+      box-shadow: 0 0 0 10px rgba($--kyc-normal-blue, 0.3);
+
+      span,
+      .material-icons {
+        color: white;
+      }
+    }
+
+    &:focus {
+      box-shadow: 0 0 0 10px rgba($--kyc-normal-blue, 0.5);
+    }
+
+    &:active {
+      box-shadow: 0 0 0 10px rgba($--kyc-normal-blue, 0.7);
+    }
+
+    &:focus-visible {
+      outline: none;
+      border: none;
+    }
+  }
+
+  &.blue {
+    @keyframes hover {
+      from {
+        background: $--kyc-normal-blue;
+        box-shadow: 0 0 1pt 0 $--kyc-normal-blue-35;
+      }
+
+      to {
+        background: $--kyc-green;
+        box-shadow: 0 0 0 1pt $--kyc-green-35;
+      }
+    }
+
+    @keyframes textHover {
+      from {
+        margin-left: 0.5em;
+      }
+
+      to {
+        margin-left: 0.2em;
+      }
+    }
+
+    background: $--kyc-normal-blue;
+    box-shadow: 0 0 1pt 0 $--kyc-normal-blue-35;
+
+    span,
+    .material-icons {
+      font-weight: 600;
+      color: white;
+      margin-left: 0.5em;
+    }
+
+    &:hover {
+      animation-name: hover;
+      animation-duration: 500ms;
+
+      background: $--kyc-green;
+      box-shadow: 0 0 0 1pt $--kyc-green-35;
+      
+      span,
+      .material-icons {
+        animation-name: textHover;
+        animation-duration: 500ms;
+
+        margin-left: 0.2em;
+        color: white;
+      }
+    }
+
+    &:focus {
+      box-shadow: 0 0 0 1pt $--kyc-green-50;
+    }
+
+    &:active {
+      box-shadow: 0 0 0 1pt $--kyc-green-75;
+    }
+
+    &:focus-visible {
+      outline: none;
+      border: none;
+    }
+  }
+
+  &.dark {
+    background: $--kyc-dark-blue;
+    box-shadow: 0 0 0 0 rgba(white, 0);
+
+    span,
+    .material-icons {
+      color: white;
+    }
+
+    &:hover {
+      background: white;
+      box-shadow: 0 0 0 10px rgba(white, 0.3);
+
+      span,
+      .material-icons {
+        color: $--kyc-normal-blue;
+      }
+    }
+
+    &:focus {
+      box-shadow: 0 0 0 10px rgba($--kyc-dark-blue, 0.5);
+    }
+
+    &:active {
+      box-shadow: 0 0 0 10px rgba($--kyc-dark-blue, 0.7);
+    }
+
+    &:focus-visible {
+      outline: none;
+      border: none;
+    }
+  }
+
+  &.passed {
+    background: $--kyc-green;
+    box-shadow: 0 0 0 0 rgba(white, 0);
+
+    span,
+    .material-icons {
+      color: $--kyc-dark-blue;
+    }
+
+    &:hover {
+      background: $--kyc-green;
+      box-shadow: 0 0 0 10px rgba(white, 0.3);
+    }
+  }
+
+  &.underline {
+    color: $--kyc-normal-blue;
+    text-decoration: underline;
+    background: transparent;
+    
+    @keyframes underlineHover {
+      from {
+        color: $--kyc-normal-blue;
+      }
+
+      to {
+        color: $--kyc-green;
+      }
+    }
+
+    &:hover {
+      animation-name: underlineHover;
+      animation-duration: 100ms;
+      color: $--kyc-green;
+    }
+  }
+}
