+++ sdk-app/src/pages/emailDiscordVerificationStep.tsx
@@ -1,9 +1,8 @@
-import { useContext, useCallback } from "react"
-import { Button } from "./button/button"
-import { Input } from "./input/input.component"
-import { StepID, DataActionTypes } from "./reducer"
-import { StateContext } from "./stateContext"
-import { Step } from "./step"
+import { useContext, useCallback, useMemo } from "react"
+import { Input } from "../components/input/input.component"
+import { StateContext, StepID, DataActionTypes } from "../components/stateContext"
+import { Step } from "../components/step"
+import { SubmitButton } from "../components/submitButton/submitButton"
 
 const emailRegex = /^[^@]+@[a-z0-9-]+.[a-z]+$/
 
@@ -17,12 +16,18 @@ export const EmailDiscordVerificationStep = () => {
     const onEmailChange = useCallback((value: string) => {
         dispatch({ payload: value, type: DataActionTypes.emailChange })
     }, [])
+    
+    const disableSubmit = useMemo(() => {
+        return !emailRegex.test(data.email)
+    }, [data.email])
 
     const onSubmit = useCallback(() => {
-        dispatch({ type: DataActionTypes.nexPage, payload: StepID.taxResidenceStep })
-    }, [])
+        if(!disableSubmit) {
+            dispatch({ type: DataActionTypes.nexPage, payload: StepID.taxResidenceStep })
+        }
+    }, [disableSubmit])
 
-    return <Step prev={onPrev} footer={
+    return <Step onEnter={ onSubmit } prev={onPrev} footer={
         <>
             {/*<span style={{ display: 'inline-flex' }}>
                 <div style={{ background: '#7289d9', width: 60, height: 60, display: "flex", justifyContent: 'center', alignItems: 'center' }}>
@@ -30,8 +35,8 @@ export const EmailDiscordVerificationStep = () => {
                 </div>
                 <p className="p" style={{ alignSelf: 'center', marginLeft: '1em' }}>Connect Discord</p>
     </span>*/}
-            <Input value={data.email} placeholder={"email"} className="full-width" onChange={onEmailChange} />
-            <Button disabled={!emailRegex.test(data.email)} className="full-width blue" onClick={onSubmit} />
+            <Input autoFocus value={data.email} placeholder={"email"} className="full-width" onChange={onEmailChange} />
+            <SubmitButton disabled={disableSubmit} className="full-width blue" onClick={onSubmit} />
         </>
     }>
         <h1 className="h1">Email / Discord verification</h1>
