+++ sdk-app/src/pages/finalStep.tsx
@@ -1,8 +1,8 @@
 import { useContext, useCallback, useState, useEffect } from "react"
 import { Step } from "../components/step"
 import { Button } from "../components/button/button"
-import { Placeholder } from "./placeholder/placeholder"
-import { KycDaoContext } from "./kycDao.provider"
+import { KycDaoContext } from "../components/kycDao.provider"
+import { Placeholder } from "../components/placeholder/placeholder"
 
 export const FinalStep = () => {
     const kycDao = useContext(KycDaoContext)
@@ -26,6 +26,6 @@ export const FinalStep = () => {
         <div style={{ display: 'flex', justifyContent: 'center' }}>
             {nftImageUrl ? <img src={nftImageUrl} width="300px" height="300px" /> : <Placeholder width="300px" height="300px" />}
         </div>
-        <Button hideArrow label="Check on chain" hoverLabel="Check on chain" className="full-width underline centered" onClick={onCheck} />
+        <Button label="Check on chain" className="full-width underline centered" onClick={onCheck} />
     </Step>
 }
