+++ sdk-app/src/pages/membershipStep.tsx
@@ -1,8 +1,8 @@
 import { FC, useContext, useCallback } from "react"
-import { Button } from "./button/button"
-import { StepID, DataActionTypes } from "./reducer"
-import { StateContext } from "./stateContext"
-import { Step } from "./step"
+import { DataActionTypes, StateContext, StepID } from "../components/stateContext"
+import { Step } from "../components/step"
+import { SubmitButton } from "../components/submitButton/submitButton"
+
 
 export const KycDAOMembershipStep: FC = () => {
     const { dispatch } = useContext(StateContext)
@@ -16,10 +16,10 @@ export const KycDAOMembershipStep: FC = () => {
         dispatch({ type: DataActionTypes.termsAcceptedChange, payload: true })
     }, [])
 
-    return <Step prev={onPrev} footer={
+    return <Step onEnter={onSubmit} prev={onPrev} footer={
         <>
             <div className="policy">By starting verification you accept <a href="#">Privacy Policy</a> and <a href="#">Terms &#38; Conditions.</a></div>
-            <Button className="full-width blue" onClick={onSubmit}/>
+            <SubmitButton className="full-width blue" onClick={onSubmit}/>
         </>
     }>
         <h1 className="h1">01 KycDAO Membership</h1>
