--- sdk-app/src/pages/steps.tsx
@@ -1,18 +1,0 @@
-import { Step } from "../components/step"
-import { Button } from "../components/button/button"
-
-export const kycNFTArtSelectionStep = () => {
-    return <Step >
-        <h1>KycDAO</h1>
-        <p>Lorem ipsum</p>
-        <Button />
-    </Step>
-}
-
-export const LastStep = () => {
-    return <Step>
-        <h1>KycDAO</h1>
-        <p>Lorem ipsum</p>
-        <Button />
-    </Step>
-}
